Update import paths after math-editor restructure (#1115)
