+++ projects/common/modules/math-editor/math-editor.module.ts
@@ -1,7 +1,7 @@
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MatButtonToggleModule } from '@angular/material/button-toggle';
-import { MathInputComponent } from 'common/modules/math-editor/math-input.component';
+import { MathInputComponent } from 'common/modules/math-editor/components/math-input/math-input.component';
 
 @NgModule({
   declarations: [
